factor out saveApiConversationHistory
